requirements.txt
pyalsaaudio
python-osc
